+++ packages/invoices/src/InboundShipment/DetailView/AppBarButtons.tsx
@@ -10,7 +10,7 @@ import {
   ReportCategory,
   LoadingButton,
 } from '@openmsupply-client/common';
-import { useInbound, useIsInboundDisabled } from '../api';
+import { useInbound } from '../api';
 import {
   ReportRowFragment,
   ReportSelector,
@@ -24,8 +24,8 @@ interface AppBarButtonProps {
 export const AppBarButtonsComponent: FC<AppBarButtonProps> = ({
   onAddItem,
 }) => {
-  const isDisabled = useIsInboundDisabled();
-  const { data } = useInbound();
+  const isDisabled = useInbound.utils.isDisabled();
+  const { data } = useInbound.document.get();
   const { OpenButton } = useDetailPanel();
   const t = useTranslation('common');
   const { print, isPrinting } = useReport.utils.print();
