+++ packages/invoices/src/InboundShipment/api/hooks/document/index.ts
@@ -1,0 +1,19 @@
+import { useDeleteSelectedInbounds } from './useDeleteSelectedInbounds';
+import { useInbound } from './useInbound';
+import { useInboundFields } from './useInboundFields';
+import { useInbounds } from './useInbounds';
+import { useInboundsAll } from './useInboundsAll';
+import { useInsertInbound } from './useInsertInbound';
+import { useNextItem } from './useNextItem';
+import { useUpdateInbound } from './useUpdateInbound';
+
+export const Document = {
+  useDeleteSelectedInbounds,
+  useInbound,
+  useInboundFields,
+  useInbounds,
+  useInboundsAll,
+  useInsertInbound,
+  useNextItem,
+  useUpdateInbound,
+};
