+++ packages/invoices/src/InboundShipment/api/hooks/document/useUpdateInbound.ts
@@ -1,0 +1,10 @@
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
+import { useInboundApi } from '../utils/useInboundApi';
+
+export const useUpdateInbound = () => {
+  const queryClient = useQueryClient();
+  const api = useInboundApi();
+  return useMutation(api.update, {
+    onSuccess: () => queryClient.invalidateQueries(api.keys.base()),
+  });
+};
