+++ packages/invoices/src/InboundShipment/api/hooks/index.ts
@@ -1,0 +1,33 @@
+import { Document } from './document';
+import { Lines } from './line';
+import { Utils } from './utils';
+
+export const useInbound = {
+  document: {
+    get: Document.useInbound,
+    list: Document.useInbounds,
+    listAll: Document.useInboundsAll,
+
+    delete: Document.useDeleteSelectedInbounds,
+    insert: Document.useInsertInbound,
+    update: Document.useUpdateInbound,
+
+    fields: Document.useInboundFields,
+    next: Document.useNextItem,
+  },
+  lines: {
+    list: Lines.useInboundLines,
+    items: Lines.useInboundItems,
+    rows: Lines.useInboundRows,
+    serviceLines: Lines.useInboundServiceLines,
+
+    delete: Lines.useDeleteInboundLines,
+    deleteSelected: Lines.useDeleteSelectedLines,
+    save: Lines.useSaveInboundLines,
+  },
+  utils: {
+    api: Utils.useInboundApi,
+    isDisabled: Utils.useIsInboundDisabled,
+    isStatusChangeDisabled: Utils.useIsStatusChangeDisabled,
+  },
+};
