+++ packages/invoices/src/InboundShipment/api/hooks/utils/index.ts
@@ -1,0 +1,9 @@
+import { useInboundApi } from './useInboundApi';
+import { useIsInboundDisabled } from './useIsInboundDisabled';
+import { useIsStatusChangeDisabled } from './useIsStatusChangeDisabled';
+
+export const Utils = {
+  useInboundApi,
+  useIsInboundDisabled,
+  useIsStatusChangeDisabled,
+};
